Fix : 위도, 경도 좌표변환 문제 해결
- float 에서 double로 필드 속성 변경
- 쿼리스트링 순서 변경
기존 latitude, longitude
변경 longitude, latitude

+++ BE/src/main/java/com/codesquad/dust6/domain/CoordinateDTO.java
@@ -1,27 +1,27 @@
 package com.codesquad.dust6.domain;
 
 public class CoordinateDTO {
-    private float longitude;
-    private float latitude;
+    private double longitude;
+    private double latitude;
 
-    public CoordinateDTO(float longitude, float latitude) {
+    public CoordinateDTO(double longitude, double latitude) {
         this.longitude = longitude;
         this.latitude = latitude;
     }
 
-    public float getLongitude() {
+    public double getLongitude() {
         return longitude;
     }
 
-    public void setLongitude(float longitude) {
+    public void setLongitude(double longitude) {
         this.longitude = longitude;
     }
 
-    public float getLatitude() {
+    public double getLatitude() {
         return latitude;
     }
 
-    public void setLatitude(float latitude) {
+    public void setLatitude(double latitude) {
         this.latitude = latitude;
     }
